Renamed block packages to blockentity, moved entity models into entity model package, deleted unused models
