+++ src/main/java/com/mraof/minestuck/client/ClientProxy.java
@@ -4,9 +4,9 @@ import com.mraof.minestuck.Minestuck;
 import com.mraof.minestuck.block.MSBlocks;
 import com.mraof.minestuck.blockentity.MSBlockEntityTypes;
 import com.mraof.minestuck.client.gui.MSScreenFactories;
-import com.mraof.minestuck.client.model.BishopModel;
+import com.mraof.minestuck.client.model.entity.BishopModel;
 import com.mraof.minestuck.client.model.MSModelLayers;
-import com.mraof.minestuck.client.model.RookModel;
+import com.mraof.minestuck.client.model.entity.RookModel;
 import com.mraof.minestuck.client.model.armor.ArmorModels;
 import com.mraof.minestuck.client.model.armor.CrumplyHatModel;
 import com.mraof.minestuck.client.model.armor.DreamerPajamasModel;
