+++ src/main/java/com/mraof/minestuck/client/model/entity/FrogModel.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.client.model;
+package com.mraof.minestuck.client.model.entity;
 
 import com.mraof.minestuck.entity.FrogEntity;
 import net.minecraft.client.model.HierarchicalModel;
