+++ src/main/resources/assets/minestuck/animations/blockentity/alchemiter.animation.json
