+++ src/main/resources/assets/minestuck/animations/blockentity/totemlathe.animation.json
