+++ src/main/resources/assets/minestuck/geo/blockentity/alchemiter.geo.json
