+++ src/main/resources/assets/minestuck/geo/blockentity/totemlathe.geo.json
